actualización

+++ QRY_DetReFact.sql
@@ -16,7 +16,7 @@ RETURNING
  CHAR(2),
  INT,
  CHAR(4), 
- SMALLINT,
+ INT,
  CHAR(1),
  INT,
  DECIMAL,
@@ -35,7 +35,7 @@ DEFINE vfolio 	INT;
 DEFINE vserie   CHAR(4);
 DEFINE vcia     CHAR(2);
 DEFINE vpla     CHAR(2);
-DEFINE vmov   	SMALLINT;
+DEFINE vmov   	INT;
 DEFINE vtipo    CHAR(1);
 DEFINE vfnvta   INT;
 DEFINE vffis    DECIMAL;
